docs: remove some examples
