--- example/src/TopBarIconTextExample.js
@@ -1,94 +1,0 @@
-import React, { Component } from 'react';
-import { View, Image, StyleSheet } from 'react-native';
-import { TabViewAnimated, TabBarTop } from 'react-native-tab-view';
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  tabbar: {
-    backgroundColor: '#222',
-  },
-  tab: {
-    padding: 4,
-  },
-  page: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  indicator: {
-    backgroundColor: '#ffeb3b',
-  },
-});
-
-export default class TopBarIconExample extends Component {
-  static propTypes = {
-    style: View.propTypes.style,
-  };
-
-  state = {
-    index: 0,
-    routes: [
-      { key: '1', title: 'First' },
-      { key: '2', title: 'Second' },
-      { key: '3', title: 'Third' },
-    ],
-  };
-
-  _handleChangeTab = (index) => {
-    this.setState({
-      index,
-    });
-  };
-
-  _renderIcon = ({ route }) => {
-    switch (route.key) {
-    case '1':
-      return <Image source={require('../assets/tab-icon-1.png')} />;
-    case '2':
-      return <Image source={require('../assets/tab-icon-2.png')} />;
-    case '3':
-      return <Image source={require('../assets/tab-icon-3.png')} />;
-    default:
-      return null;
-    }
-  };
-
-  _renderHeader = (props) => {
-    return (
-      <TabBarTop
-        {...props}
-        indicatorStyle={styles.indicator}
-        renderIcon={this._renderIcon}
-        tabStyle={styles.tab}
-        style={styles.tabbar}
-      />
-    );
-  };
-
-  _renderScene = ({ route }) => {
-    switch (route.key) {
-    case '1':
-      return <View style={[ styles.page, { backgroundColor: '#ff4081' } ]} />;
-    case '2':
-      return <View style={[ styles.page, { backgroundColor: '#673ab7' } ]} />;
-    case '3':
-      return <View style={[ styles.page, { backgroundColor: '#4caf50' } ]} />;
-    default:
-      return null;
-    }
-  };
-
-  render() {
-    return (
-      <TabViewAnimated
-        style={[ styles.container, this.props.style ]}
-        navigationState={this.state}
-        renderScene={this._renderScene}
-        renderHeader={this._renderHeader}
-        onRequestChangeTab={this._handleChangeTab}
-      />
-    );
-  }
-}
